Move flask app configuration outside python package
Move configuration outside the package hierarchy. Makes it a bit easier
to find. Also, when coupled with the fact that the config root path can
be provided in the environment, this makes it possible to produce a
source package that is not tied to a given environment.

Make configuration hierarchical. This allows common configuration to be
set in a root config while retaining the ability to override it in a
specific app.

Set the GOVUK_ENV environment variable to test in unittests (this is
taken from @alexmuller's commit e1224c4)

This commit is backwards compatible, the legacy config location is used
as that is still where alphagov-deployment is putting it.